requirements.txt
imageio==2.6.1
joblib==0.14.1
matplotlib==3.1.2
nibabel==3.0.0
nilearn==0.6.0
numpy==1.18.1
pandas==0.25.3
scikit-learn==0.22.1
scipy==1.4.1
sklearn-rvm==0.1
tqdm==4.41.1
statsmodels==0.10.2